moved file paths to models settings json
It seems that if the current settings file has a field
for whether the model is to be saved, it can also have
a field for where on disk the output should go. This was
added as "savedir" (in a backward compatible way for now).

Similarly the input locations and field name for the hdf5
files were moved to the external settings file.