added argparse and image outputs to sample.py
Added ArguementParser processing so that more complicated
arguments could be passed on the command line. One
argument added was "weights", which overrides the
"pathsToWeights" setting; the rationale for adding this
was that a single model.json file might have multiple
checkpointed outputs, so it's probably easier to be able
to specify the file directly on the command line.

Added a second mode for samples, which is to output
a series of files to a subdirectory. I'm guessing the
pylab code that was there was meant to be run in the
context of ipython (not sure), so for now I'm switching
modes based on whether ipython is running.

Code also added to create an animated gif at the end
showing the full sequence. Note that this adds a dependency
on the imagemagick convert command line tool.